Add file node-maker.js. Remove get-comment.js, get-thumbnail.js.
